requirements.txt
aniso8601==9.0.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
fastai==2.2.7
fastcore==1.3.19
fastprogress==1.0.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.0
murmurhash==1.0.5
numpy==1.20.2
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
plac==1.1.3
preshed==3.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
